LayoutTests/imported/w3c/web-platform-tests/css/css-view-transitions/web-animation-pseudo-incorrect-name-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: SyntaxError: The string did not match the expected pattern.

Harness Error (FAIL), message = Unhandled rejection: The string did not match the expected pattern.

TIMEOUT animation created with incorrect name Test timed out

Harness Error (FAIL), message = Unhandled rejection: The string did not match the expected pattern.

TIMEOUT animation created with incorrect name Test timed out
